+++ src/Utils.ts
@@ -1,22 +1,12 @@
-interface DisplayTime {
-    minutes: string
-    seconds: string,
-    milliSeconds: string
-}
-
-export const formatTime = (time: number): DisplayTime => {
-    const minutes = Math.floor(time / (1000 * 60))
-    const seconds = Math.floor((time % (1000 * 60)) / 1000)
-    const milliseconds = Math.floor((time % 1000) / 10)
+export const formatTime = (time: number): string => {
+    const minutes = Math.floor(time / (1000 * 60));
+    const seconds = Math.floor((time % (1000 * 60)) / 1000);
+    const milliseconds = Math.floor((time % 1000) / 10);
 
     // Pad single-digit seconds and milliseconds with leading zeros
-    const formattedMinutes = minutes < 10 ? `0${minutes}` : `${minutes}`
-    const formattedSeconds = seconds < 10 ? `0${seconds}` : `${seconds}`
-    const formattedMilliseconds = milliseconds < 10 ? `0${milliseconds}` : `${milliseconds}`
+    const formattedMinutes = minutes < 10 ? `0${minutes}` : `${minutes}`;
+    const formattedSeconds = seconds < 10 ? `0${seconds}` : `${seconds}`;
+    const formattedMilliseconds = milliseconds < 10 ? `0${milliseconds}` : `${milliseconds}`;
 
-    return {
-        minutes: formattedMinutes,
-        seconds: formattedSeconds,
-        milliSeconds: formattedMilliseconds
-    }
-}
+    return `${formattedMinutes}:${formattedSeconds}:${formattedMilliseconds}`;
+};
