+++ src/components/LapGroup.tsx
@@ -1,8 +1,8 @@
-import {Animated, StyleSheet, View} from "react-native";
+import {Animated, StyleSheet} from "react-native";
 import Lap, {LapProps} from "./Lap";
 import React from "react";
-import FlatList = Animated.FlatList;
 import LapSeparator from "./LapSeparator";
+import FlatList = Animated.FlatList;
 
 interface LapGroupProps {
     laps: LapProps[]
@@ -18,6 +18,7 @@ export default function LapGroup({laps}: LapGroupProps) {
             style={styles.container}
             ListHeaderComponent={LapSeparator}
             ListFooterComponent={LapSeparator}
+            testID={"lap-list"}
         />
     )
